requirements.txt
psycopg[binary]
pyyaml